:truck: move path handling functions to their own script
